Move the attribute filter logic to shared_filters file
